requirements.txt
flask==3.1.3
anthropic==0.84.0
python-dotenv==1.2.1
gunicorn==23.0.0
